Complete 3 class ( Register + Login + My Account)

+++ actions/commons/GlobalConstants.java
@@ -13,9 +13,9 @@ public class GlobalConstants {
 		public static final String DRAG_DROP_HTML5 = PROJECT_PATH + File.separator + "drapDropHTML5";
 		public static final String REPORTNG_SCREENSHOT = PROJECT_PATH + File.separator + "reportNGImages" + File.separator;
 		
-		public static final int LONG_TIME_OUT = 10;
-		public static final int SHORT_TIME_OUT = 5;
-		public static final int RETRY_TEST_FAIL = 3;
+		public static final long LONG_TIME_OUT = 10;
+		public static final long SHORT_TIME_OUT = 5;
+		public static final long RETRY_TEST_FAIL = 3;
 		
 		
 		public static final String LIVE_TECHPANDA_USER = "http://live.techpanda.org/index.php/";
